32 bit add

+++ addition.c
@@ -1,12 +1,12 @@
 #include <stdio.h>
 #include <stdint.h>
 
-volatile uint8_t c;
-uint8_t add( uint8_t a, uint8_t b) {
+volatile uint32_t c;
+uint32_t add( uint32_t a, uint32_t b) {
 return a + b;
 }
-int main( void) {
-uint8_t a = 1, b = 2;
+uint32_t main( void) {
+uint32_t a = 1, b = 2;
 c = add( a, b);
 printf( "a + b = %i\n", c);
 exit( 1);
